+++ python/pyspark/shuffle.py
@@ -28,8 +28,7 @@ import sys
 import heapq
 from pyspark.serializers import BatchedSerializer, PickleSerializer, FlattenedValuesSerializer, \
     CompressedSerializer, AutoBatchedSerializer
-from pyspark.util import fail_on_stopiteration
-from pyspark.warnings import PySparkWarning
+from pyspark.util import fail_on_stopiteration, PySparkWarning
 
 
 try:
