+++ python/pyspark/sql/session.py
@@ -31,7 +31,7 @@ from pyspark.sql.types import DataType, StructType, \
     _make_type_verifier, _infer_schema, _has_nulltype, _merge_type, _create_converter, \
     _parse_datatype_string
 from pyspark.sql.utils import install_exception_handler
-from pyspark.warnings import PySparkWarning
+from pyspark.util import PySparkWarning
 
 __all__ = ["SparkSession"]
